pyproject.toml
[project]
name = "pdf-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.120.1",
    "hf-xet>=1.2.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.1",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.0",
    "pymupdf>=1.26.5",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
    "structlog>=25.5.0",
    "uvicorn>=0.38.0",
]
